Creating main.go
Added functionality to the main.go to enable a welcome page to webpage.
This is backed up with a html page set and css formatting. This test is
to prove that the inital code works this will be developed from this
point to test adding more functionality onto the webpage

files added
main.go
welcome-template.css
welcome-template.html

This displays a simple welcome page showing a welcome message with the
time and date