Add Manifest files
Adding files to create a namespace and deploy database, an application
and a service within that namespace. This will allow for a application
to run within a kube cluster with a single database instance.

sql contains two sql scripts that will create a table within postgres
and the second that will populate the table with data that is used
within the application.